# Performance regression in stat_ops.FrameMultiIndexOps.time_op