@simonjayhawkins 2020-07-22
@jreback if we're not branching, should I reopen this and merge to master? see https://github.com/pandas-dev/pandas/issues/34730#issuecomment-656182086